Refactor: split escrow and oracle test files into thematic modules
Escrow (#655):
- Deleted contracts/escrow/src/tests.rs (single flat file, ~3100 lines)
- Created contracts/escrow/src/tests/mod.rs — shared setup(), mint_player_balance() helpers and module declarations
- Created contracts/escrow/src/tests/lifecycle.rs — match creation, deposit, payout, cancel, and expiry tests
- Created contracts/escrow/src/tests/admin.rs — pause/unpause, oracle rotation, admin transfer, and auth tests
- Created contracts/escrow/src/tests/events.rs — event emission verification tests
- Created contracts/escrow/src/tests/ttl.rs — TTL extension tests for every storage-touching operation

Oracle (#656):
- Extracted the inline #[cfg(test)] mod tests { … } block from contracts/oracle/src/lib.rs into a standalone contracts/oracle/src/tests.rs
- Replaced the inline block with a single #[cfg(test)] mod tests; declaration in lib.rs
- Preserved all imports, the setup() helper, and every test function verbatim

Shared helpers (setup, mint_player_balance) live in tests/mod.rs and are re-exported to all escrow submodules via pub use. Each submodule brings in its own trait imports (Address as _, Ledger as _, Persistent as _, etc.) to keep inter-module coupling minimal.

Closes #655
Closes #656
