--- contracts/escrow/src/tests.rs
@@ -1,3103 +1,0 @@
-extern crate std;
-
-use super::*;
-use soroban_sdk::{
-    testutils::{
-        storage::{Instance as _, Persistent as _},
-        Address as _, Events, Ledger as _, MockAuth, MockAuthInvoke,
-    },
-    token::{Client as TokenClient, StellarAssetClient},
-    vec, Address, Env, IntoVal, String, Symbol, TryFromVal,
-};
-
-fn setup() -> (Env, Address, Address, Address, Address, Address, Address) {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let oracle = Address::generate(&env);
-    let player1 = Address::generate(&env);
-    let player2 = Address::generate(&env);
-
-    let token_id = env.register_stellar_asset_contract_v2(admin.clone());
-    let token_addr = token_id.address();
-    let asset_client = StellarAssetClient::new(&env, &token_addr);
-    asset_client.mint(&player1, &1000);
-    asset_client.mint(&player2, &1000);
-
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-    client.initialize(&oracle, &admin);
-
-    (
-        env,
-        contract_id,
-        oracle,
-        player1,
-        player2,
-        token_addr,
-        admin,
-    )
-}
-
-fn mint_player_balance(asset_client: &StellarAssetClient, player: &Address, amount: i128) {
-    asset_client.mint(player, &amount);
-}
-
-#[test]
-fn test_initialize_emits_event() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let oracle = Address::generate(&env);
-
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-    client.initialize(&oracle, &admin);
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "escrow").into_val(&env),
-        symbol_short!("init").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "escrow initialized event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let (ev_oracle, ev_admin): (Address, Address) = TryFromVal::try_from_val(&env, &data).unwrap();
-    assert_eq!(ev_oracle, oracle);
-    assert_eq!(ev_admin, admin);
-}
-
-#[test]
-fn test_is_initialized_false_before_initialize_and_true_after() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let oracle = Address::generate(&env);
-
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    assert!(
-        !client.is_initialized(),
-        "contract must report uninitialized before initialize is called"
-    );
-
-    client.initialize(&oracle, &admin);
-
-    assert!(
-        client.is_initialized(),
-        "contract must report initialized after initialize is called"
-    );
-}
-
-#[test]
-fn test_create_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "abc123"),
-        &Platform::Lichess,
-    );
-
-    assert_eq!(id, 0);
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Pending);
-}
-
-#[test]
-fn test_match_state_pending_immediately_after_create_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "pending_state_test"),
-        &Platform::Lichess,
-    );
-
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Pending);
-    assert!(!m.player1_deposited);
-    assert!(!m.player2_deposited);
-}
-
-#[test]
-fn test_get_match_returns_stake_and_token() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let stake_amount = 500i128;
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &stake_amount,
-        &token,
-        &String::from_str(&env, "game_266"),
-        &Platform::Lichess,
-    );
-
-    let m = client.get_match(&id);
-    assert_eq!(m.stake_amount, stake_amount);
-    assert_eq!(m.token, token);
-}
-
-#[test]
-fn test_deposit_and_activate() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "abc123"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    assert!(!client.is_funded(&id));
-    client.deposit(&id, &player2);
-    assert!(client.is_funded(&id));
-    assert_eq!(client.get_escrow_balance(&id), 200);
-}
-
-#[test]
-fn test_concurrent_deposits_same_ledger() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "concurrent_deposits"),
-        &Platform::Lichess,
-    );
-
-    // Deposits back-to-back without ledger advancement
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Active);
-    assert!(client.is_funded(&id));
-}
-
-#[test]
-fn test_is_funded_false_after_only_player1_deposits() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "partial_funded_game"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    assert!(
-        !client.is_funded(&id),
-        "is_funded must be false after only player1 deposits"
-    );
-
-    client.deposit(&id, &player2);
-    assert!(
-        client.is_funded(&id),
-        "is_funded must be true after both players deposit"
-    );
-}
-
-/// Verify the deposit flags on the Match struct after each individual deposit.
-#[test]
-fn test_deposit_flags_set_correctly_after_each_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "deposit_flags_test"),
-        &Platform::Lichess,
-    );
-
-    let m = client.get_match(&id);
-    assert!(
-        !m.player1_deposited,
-        "player1_deposited must be false before any deposit"
-    );
-    assert!(
-        !m.player2_deposited,
-        "player2_deposited must be false before any deposit"
-    );
-
-    client.deposit(&id, &player1);
-    let m = client.get_match(&id);
-    assert!(
-        m.player1_deposited,
-        "player1_deposited must be true after player1 deposits"
-    );
-    assert!(
-        !m.player2_deposited,
-        "player2_deposited must still be false after only player1 deposits"
-    );
-
-    client.deposit(&id, &player2);
-    let m = client.get_match(&id);
-    assert!(
-        m.player1_deposited,
-        "player1_deposited must remain true after player2 deposits"
-    );
-    assert!(
-        m.player2_deposited,
-        "player2_deposited must be true after player2 deposits"
-    );
-}
-
-#[test]
-fn test_full_match_lifecycle_winner_and_draw_scenarios() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-    let asset_client = StellarAssetClient::new(&env, &token);
-    let player3 = Address::generate(&env);
-    let player4 = Address::generate(&env);
-
-    mint_player_balance(&asset_client, &player3, 1000);
-    mint_player_balance(&asset_client, &player4, 1000);
-
-    let winner_match_id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "full_lifecycle_winner"),
-        &Platform::Lichess,
-    );
-
-    let winner_match = client.get_match(&winner_match_id);
-    assert_eq!(winner_match.state, MatchState::Pending);
-    assert_eq!(token_client.balance(&player1), 1000);
-    assert_eq!(token_client.balance(&player2), 1000);
-    assert_eq!(client.get_escrow_balance(&winner_match_id), 0);
-
-    client.deposit(&winner_match_id, &player1);
-    let winner_match = client.get_match(&winner_match_id);
-    assert_eq!(winner_match.state, MatchState::Pending);
-    assert!(winner_match.player1_deposited);
-    assert!(!winner_match.player2_deposited);
-    assert_eq!(token_client.balance(&player1), 900);
-    assert_eq!(token_client.balance(&player2), 1000);
-    assert_eq!(client.get_escrow_balance(&winner_match_id), 100);
-
-    client.deposit(&winner_match_id, &player2);
-    let winner_match = client.get_match(&winner_match_id);
-    assert_eq!(winner_match.state, MatchState::Active);
-    assert!(winner_match.player1_deposited);
-    assert!(winner_match.player2_deposited);
-    assert_eq!(token_client.balance(&player1), 900);
-    assert_eq!(token_client.balance(&player2), 900);
-    assert_eq!(client.get_escrow_balance(&winner_match_id), 200);
-
-    client.submit_result(&winner_match_id, &Winner::Player1);
-    let winner_match = client.get_match(&winner_match_id);
-    assert_eq!(winner_match.state, MatchState::Completed);
-    assert_eq!(token_client.balance(&player1), 1100);
-    assert_eq!(token_client.balance(&player2), 900);
-    assert_eq!(client.get_escrow_balance(&winner_match_id), 0);
-
-    let draw_match_id = client.create_match(
-        &player3,
-        &player4,
-        &75,
-        &token,
-        &String::from_str(&env, "full_lifecycle_draw"),
-        &Platform::ChessDotCom,
-    );
-
-    let draw_match = client.get_match(&draw_match_id);
-    assert_eq!(draw_match.state, MatchState::Pending);
-    assert_eq!(token_client.balance(&player3), 1000);
-    assert_eq!(token_client.balance(&player4), 1000);
-    assert_eq!(client.get_escrow_balance(&draw_match_id), 0);
-
-    client.deposit(&draw_match_id, &player3);
-    let draw_match = client.get_match(&draw_match_id);
-    assert_eq!(draw_match.state, MatchState::Pending);
-    assert_eq!(token_client.balance(&player3), 925);
-    assert_eq!(token_client.balance(&player4), 1000);
-    assert_eq!(client.get_escrow_balance(&draw_match_id), 75);
-
-    client.deposit(&draw_match_id, &player4);
-    let draw_match = client.get_match(&draw_match_id);
-    assert_eq!(draw_match.state, MatchState::Active);
-    assert_eq!(token_client.balance(&player3), 925);
-    assert_eq!(token_client.balance(&player4), 925);
-    assert_eq!(client.get_escrow_balance(&draw_match_id), 150);
-
-    client.submit_result(&draw_match_id, &Winner::Draw);
-    let draw_match = client.get_match(&draw_match_id);
-    assert_eq!(draw_match.state, MatchState::Completed);
-    assert_eq!(token_client.balance(&player3), 1000);
-    assert_eq!(token_client.balance(&player4), 1000);
-    assert_eq!(client.get_escrow_balance(&draw_match_id), 0);
-}
-
-#[test]
-fn test_full_match_lifecycle() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    // Step 1: create_match → Pending
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "lifecycle_game"),
-        &Platform::Lichess,
-    );
-    assert_eq!(client.get_match(&id).state, MatchState::Pending);
-    assert_eq!(client.get_escrow_balance(&id), 0);
-
-    // Step 2: player1 deposits → still Pending
-    client.deposit(&id, &player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Pending);
-    assert_eq!(token_client.balance(&player1), 900);
-    assert_eq!(client.get_escrow_balance(&id), 100);
-
-    // Step 3: player2 deposits → Active
-    client.deposit(&id, &player2);
-    assert_eq!(client.get_match(&id).state, MatchState::Active);
-    assert_eq!(token_client.balance(&player2), 900);
-    assert_eq!(client.get_escrow_balance(&id), 200);
-
-    // Step 4: submit_result → Completed, winner paid, escrow zeroed
-    client.submit_result(&id, &Winner::Player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Completed);
-    assert_eq!(token_client.balance(&player1), 1100); // won the pot
-    assert_eq!(token_client.balance(&player2), 900); // lost stake
-    assert_eq!(client.get_escrow_balance(&id), 0);
-}
-
-#[test]
-fn test_payout_winner() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game1"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    client.submit_result(&id, &Winner::Player1);
-
-    // player1 started with 1000, deposited 100, won the 200 pot → 1100
-    assert_eq!(token_client.balance(&player1), 1100);
-    assert_eq!(client.get_match(&id).state, MatchState::Completed);
-    assert!(client.get_match(&id).completed_ledger.is_some());
-}
-
-#[test]
-fn test_draw_refund() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game2"),
-        &Platform::ChessDotCom,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    client.submit_result(&id, &Winner::Draw);
-
-    assert_eq!(token_client.balance(&player1), 1000);
-    assert_eq!(token_client.balance(&player2), 1000);
-}
-
-#[test]
-fn test_player2_balance_decreases_after_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "player2_balance_after_deposit"),
-        &Platform::Lichess,
-    );
-
-    let balance_before = token_client.balance(&player2);
-    client.deposit(&id, &player2);
-    let balance_after = token_client.balance(&player2);
-
-    assert_eq!(balance_before, 1000);
-    assert_eq!(balance_after, 900);
-    assert_eq!(balance_before - balance_after, 100);
-}
-
-#[test]
-fn test_cancel_refunds_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game3"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.cancel_match(&id, &player1);
-
-    assert_eq!(token_client.balance(&player1), 1000);
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-}
-
-#[test]
-fn test_create_match_emits_event() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_ev2"),
-        &Platform::Lichess,
-    );
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "match").into_val(&env),
-        soroban_sdk::symbol_short!("created").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "match created event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let (ev_id, ev_p1, ev_p2, ev_stake): (u64, Address, Address, i128) =
-        TryFromVal::try_from_val(&env, &data).unwrap();
-    assert_eq!(ev_id, id);
-    assert_eq!(ev_p1, player1);
-    assert_eq!(ev_p2, player2);
-    assert_eq!(ev_stake, 100);
-}
-
-#[test]
-fn test_submit_result_emits_event() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_evt"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    client.submit_result(&id, &Winner::Player1);
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "match").into_val(&env),
-        soroban_sdk::symbol_short!("completed").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "match completed event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let decoded: (u64, Winner) = <(u64, Winner)>::try_from_val(&env, &data).unwrap();
-    assert_eq!(decoded, (id, Winner::Player1));
-}
-
-#[test]
-fn test_submit_result_fails_if_not_fully_funded() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_nofund"),
-        &Platform::Lichess,
-    );
-
-    // Only player1 deposits — player2 has not
-    client.deposit(&id, &player1);
-
-    env.as_contract(&contract_id, || {
-        let mut m: Match = env.storage().persistent().get(&DataKey::Match(id)).unwrap();
-        m.state = MatchState::Active;
-        env.storage().persistent().set(&DataKey::Match(id), &m);
-    });
-
-    let result = client.try_submit_result(&id, &Winner::Player1);
-    assert_eq!(result, Err(Ok(Error::NotFunded)));
-}
-
-#[test]
-fn test_submit_result_fails_when_contract_token_balance_is_zero() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "zero_balance_game"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    // Drain the contract's token balance by transferring out all funds
-    let contract_balance = token_client.balance(&contract_id);
-    if contract_balance > 0 {
-        env.as_contract(&contract_id, || {
-            token_client.transfer(&contract_id, &player1, &contract_balance);
-        });
-    }
-
-    // Verify balance is zero
-    assert_eq!(token_client.balance(&contract_id), 0);
-
-    // Attempt to submit result should fail due to insufficient balance
-    let result = client.try_submit_result(&id, &Winner::Player1);
-    assert!(
-        result.is_err(),
-        "submit_result should fail when contract has zero token balance"
-    );
-}
-
-#[test]
-fn test_initialize_accepts_valid_generated_oracle_address() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle = Address::generate(&env);
-    let admin = Address::generate(&env);
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.initialize(&oracle, &admin);
-
-    let stored_oracle: Address = env.as_contract(&contract_id, || {
-        env.storage().instance().get(&DataKey::Oracle).unwrap()
-    });
-    assert_eq!(stored_oracle, oracle);
-}
-
-#[test]
-fn test_initialize_rejects_contract_address_as_oracle() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // Passing the contract's own address as oracle must be rejected
-    let result = client.try_initialize(&contract_id, &admin);
-    assert_eq!(result, Err(Ok(Error::InvalidAddress)));
-}
-
-#[test]
-fn test_cancel_match_emits_event() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_cancel"),
-        &Platform::Lichess,
-    );
-
-    client.cancel_match(&id, &player1);
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "match").into_val(&env),
-        soroban_sdk::symbol_short!("cancelled").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "match cancelled event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let ev_id: u64 = TryFromVal::try_from_val(&env, &data).unwrap();
-    assert_eq!(ev_id, id);
-}
-
-#[test]
-fn test_cancel_match_no_deposits_emits_no_token_transfers() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_no_deposit_cancel"),
-        &Platform::Lichess,
-    );
-
-    client.cancel_match(&id, &player1);
-
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-
-    // No token transfers should have been emitted — neither player deposited
-    let transfer_topic: soroban_sdk::Val = soroban_sdk::symbol_short!("transfer").into_val(&env);
-    let has_transfer = env
-        .events()
-        .all()
-        .iter()
-        .any(|(_, topics, _)| topics.contains(transfer_topic));
-    assert!(
-        !has_transfer,
-        "no token transfer events should be emitted when no deposits were made"
-    );
-}
-
-#[test]
-fn test_concurrent_matches_remain_isolated() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let oracle = Address::generate(&env);
-    let player1 = Address::generate(&env);
-    let player2 = Address::generate(&env);
-    let player3 = Address::generate(&env);
-    let player4 = Address::generate(&env);
-
-    let token_id = env.register_stellar_asset_contract_v2(admin.clone());
-    let token = token_id.address();
-    let asset_client = StellarAssetClient::new(&env, &token);
-    let token_client = TokenClient::new(&env, &token);
-
-    for player in [&player1, &player2, &player3, &player4] {
-        mint_player_balance(&asset_client, player, 1000);
-    }
-
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-    client.initialize(&oracle, &admin);
-
-    let match_one = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "concurrent_match_one"),
-        &Platform::Lichess,
-    );
-    let match_two = client.create_match(
-        &player3,
-        &player4,
-        &60,
-        &token,
-        &String::from_str(&env, "concurrent_match_two"),
-        &Platform::ChessDotCom,
-    );
-
-    client.deposit(&match_one, &player1);
-    client.deposit(&match_two, &player3);
-    assert_eq!(client.get_match(&match_one).state, MatchState::Pending);
-    assert_eq!(client.get_match(&match_two).state, MatchState::Pending);
-    assert_eq!(client.get_escrow_balance(&match_one), 100);
-    assert_eq!(client.get_escrow_balance(&match_two), 60);
-    assert_eq!(token_client.balance(&player1), 900);
-    assert_eq!(token_client.balance(&player2), 1000);
-    assert_eq!(token_client.balance(&player3), 940);
-    assert_eq!(token_client.balance(&player4), 1000);
-
-    client.deposit(&match_one, &player2);
-    client.deposit(&match_two, &player4);
-    assert_eq!(client.get_match(&match_one).state, MatchState::Active);
-    assert_eq!(client.get_match(&match_two).state, MatchState::Active);
-    assert_eq!(client.get_escrow_balance(&match_one), 200);
-    assert_eq!(client.get_escrow_balance(&match_two), 120);
-
-    client.submit_result(&match_one, &Winner::Player1);
-    client.submit_result(&match_two, &Winner::Draw);
-
-    assert_eq!(client.get_match(&match_one).state, MatchState::Completed);
-    assert_eq!(client.get_match(&match_two).state, MatchState::Completed);
-    assert_eq!(token_client.balance(&player1), 1100);
-    assert_eq!(token_client.balance(&player2), 900);
-    assert_eq!(token_client.balance(&player3), 1000);
-    assert_eq!(token_client.balance(&player4), 1000);
-}
-
-#[test]
-fn test_concurrent_matches_do_not_share_escrow_balances() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let oracle = Address::generate(&env);
-    let player1 = Address::generate(&env);
-    let player2 = Address::generate(&env);
-    let player3 = Address::generate(&env);
-    let player4 = Address::generate(&env);
-
-    let token_id = env.register_stellar_asset_contract_v2(admin.clone());
-    let token = token_id.address();
-    let asset_client = StellarAssetClient::new(&env, &token);
-
-    for player in [&player1, &player2, &player3, &player4] {
-        mint_player_balance(&asset_client, player, 1000);
-    }
-
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-    client.initialize(&oracle, &admin);
-
-    let match_a = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "isolated_balance_match_a"),
-        &Platform::Lichess,
-    );
-    let match_b = client.create_match(
-        &player3,
-        &player4,
-        &60,
-        &token,
-        &String::from_str(&env, "isolated_balance_match_b"),
-        &Platform::ChessDotCom,
-    );
-
-    client.deposit(&match_a, &player1);
-
-    assert_eq!(client.get_escrow_balance(&match_a), 100);
-    assert_eq!(client.get_escrow_balance(&match_b), 0);
-}
-
-#[test]
-#[should_panic(expected = "Contract already initialized")]
-fn test_double_initialize_fails() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let oracle1 = Address::generate(&env);
-    let oracle2 = Address::generate(&env);
-    let admin = Address::generate(&env);
-
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.initialize(&oracle1, &admin);
-    client.initialize(&oracle2, &admin);
-}
-
-#[test]
-fn test_pause_on_uninitialized_contract_returns_unauthorized() {
-    let env = Env::default();
-    env.mock_all_auths();
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // No initialize call — Admin key is absent
-    let result = client.try_pause();
-    assert_eq!(result, Err(Ok(Error::Unauthorized)));
-}
-
-#[test]
-fn test_admin_pause_blocks_create_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.pause();
-
-    let result = client.try_create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "paused_game"),
-        &Platform::Lichess,
-    );
-    assert_eq!(result, Err(Ok(Error::ContractPaused)));
-}
-
-#[test]
-fn test_admin_unpause_allows_create_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.pause();
-    client.unpause();
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "unpaused_game"),
-        &Platform::Lichess,
-    );
-    assert_eq!(id, 0);
-}
-
-#[test]
-fn test_pause_emits_paused_event() {
-    let (env, contract_id, ..) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.pause();
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "admin").into_val(&env),
-        symbol_short!("paused").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "paused event not emitted");
-}
-
-/// Test that deposit is rejected when the contract is paused.
-/// This verifies the invariant: no deposits can be made while the contract is paused,
-/// preventing players from locking funds in a paused state.
-#[test]
-fn test_paused_contract_rejects_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // Create a match before pausing
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game123"),
-        &Platform::Lichess,
-    );
-
-    // Admin pauses the contract
-    client.pause();
-
-    // Attempt to deposit - should fail with ContractPaused
-    let result = client.try_deposit(&id, &player1);
-    assert_eq!(result, Err(Ok(Error::ContractPaused)));
-}
-
-#[test]
-fn test_deposit_blocked_when_paused() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "paused_deposit_game"),
-        &Platform::Lichess,
-    );
-
-    client.pause();
-
-    let result = client.try_deposit(&id, &player1);
-    assert_eq!(
-        result,
-        Err(Ok(Error::ContractPaused)),
-        "deposit must return ContractPaused when the contract is paused"
-    );
-}
-
-#[test]
-fn test_deposit_by_unauthorized_address_returns_unauthorized() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "unauth_deposit_game"),
-        &Platform::Lichess,
-    );
-
-    // A random third-party address that is not player1 or player2
-    let unauthorized_address = Address::generate(&env);
-
-    let result = client.try_deposit(&id, &unauthorized_address);
-    assert_eq!(result, Err(Ok(Error::Unauthorized)));
-}
-
-#[test]
-fn test_submit_result_blocked_when_paused() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "paused_submit_game"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    client.pause();
-
-    let result = client.try_submit_result(&id, &Winner::Player1);
-    assert_eq!(result, Err(Ok(Error::ContractPaused)));
-}
-
-#[test]
-fn test_admin_can_rotate_oracle() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let next_oracle = Address::generate(&env);
-    client.update_oracle(&next_oracle);
-    assert_eq!(client.get_oracle(), next_oracle);
-
-    let attacker = Address::generate(&env);
-    let rotate_to = Address::generate(&env);
-
-    env.mock_auths(&[MockAuth {
-        address: &attacker,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "update_oracle",
-            args: (rotate_to.clone(),).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-
-    assert!(client.try_update_oracle(&rotate_to).is_err());
-}
-
-#[test]
-fn test_update_oracle_rejects_self_address() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_update_oracle(&contract_id);
-    assert_eq!(result, Err(Ok(Error::InvalidAddress)));
-}
-
-#[test]
-fn test_old_oracle_rejected_after_rotation() {
-    let (env, contract_id, oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let new_oracle = Address::generate(&env);
-    client.update_oracle(&new_oracle);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "oracle_rotation"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    env.mock_auths(&[MockAuth {
-        address: &oracle,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "submit_result",
-            args: (id, Winner::Player2).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-
-    let result = client.try_submit_result(&id, &Winner::Player2);
-    assert!(
-        matches!(result, Err(Err(_))),
-        "old oracle must not be able to submit results"
-    );
-
-    env.mock_auths(&[MockAuth {
-        address: &new_oracle,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "submit_result",
-            args: (id, Winner::Player2).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-
-    client.submit_result(&id, &Winner::Player2);
-    assert_eq!(client.get_match(&id).state, MatchState::Completed);
-}
-
-#[test]
-#[should_panic(expected = "Error(Contract, #10)")]
-fn test_create_match_with_zero_stake_fails() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // This should fail because stake_amount is 0
-    let _id = client.create_match(
-        &player1,
-        &player2,
-        &0,
-        &token,
-        &String::from_str(&env, "zero_stake_game"),
-        &Platform::Lichess,
-    );
-}
-
-#[test]
-fn test_create_match_with_negative_stake_returns_invalid_amount() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_create_match(
-        &player1,
-        &player2,
-        &-100,
-        &token,
-        &String::from_str(&env, "negative_stake_game"),
-        &Platform::Lichess,
-    );
-    assert_eq!(result, Err(Ok(Error::InvalidAmount)));
-}
-
-#[test]
-fn test_create_match_with_empty_game_id_returns_invalid_game_id() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, ""),
-        &Platform::Lichess,
-    );
-    assert_eq!(result, Err(Ok(Error::InvalidGameId)));
-}
-
-#[test]
-fn test_player2_cancel_pending_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_p2_cancel"),
-        &Platform::Lichess,
-    );
-
-    // Player2 cancels the pending match
-    client.cancel_match(&id, &player2);
-
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-}
-
-#[test]
-fn test_player2_cancel_refunds_both_players() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_p2_cancel_refund"),
-        &Platform::Lichess,
-    );
-
-    // Both players deposit - this changes state to Active
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    // Now the match is Active, not Pending - cancel should fail with InvalidState
-    let result = client.try_cancel_match(&id, &player2);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_player2_cancel_only_player2_deposited() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_p2_only"),
-        &Platform::Lichess,
-    );
-
-    // Only player2 deposits (player1 abandoned)
-    client.deposit(&id, &player2);
-
-    // Player2 cancels and gets refund
-    client.cancel_match(&id, &player2);
-
-    assert_eq!(token_client.balance(&player2), 1000);
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-}
-
-#[test]
-fn test_cancel_active_match_fails_with_invalid_state() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_active_cancel"),
-        &Platform::Lichess,
-    );
-
-    // Both players deposit — transitions match to Active
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    // Verify match is Active before attempting cancel
-    assert_eq!(client.get_match(&id).state, MatchState::Active);
-
-    // Attempt to cancel an Active match — must return MatchAlreadyActive (error code #11)
-    let result = client.try_cancel_match(&id, &player1);
-    assert_eq!(
-        result,
-        Err(Ok(Error::MatchAlreadyActive)),
-        "expected MatchAlreadyActive error when cancelling an Active match"
-    );
-
-    // Match must still be Active — no state change
-    assert_eq!(client.get_match(&id).state, MatchState::Active);
-
-    // Funds must remain in escrow — balances unchanged from post-deposit state
-    assert_eq!(token_client.balance(&player1), 900);
-    assert_eq!(token_client.balance(&player2), 900);
-}
-
-#[test]
-fn test_cancel_active_match_returns_match_already_active() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_already_active"),
-        &Platform::Lichess,
-    );
-
-    // Fund both players — match transitions to Active
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    assert_eq!(client.get_match(&id).state, MatchState::Active);
-
-    // cancel_match must return MatchAlreadyActive, not InvalidState
-    let result = client.try_cancel_match(&id, &player1);
-    assert_eq!(result, Err(Ok(Error::MatchAlreadyActive)));
-}
-
-#[test]
-#[should_panic(expected = "Error(Contract, #4)")]
-fn test_unauthorized_player_cannot_cancel() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "game_unauthorized"),
-        &Platform::Lichess,
-    );
-
-    // Create a third party who is not part of the match
-    let unauthorized = Address::generate(&env);
-
-    // This should panic with Unauthorized error
-    client.cancel_match(&id, &unauthorized);
-}
-
-#[test]
-fn test_cancel_match_on_cancelled_match_returns_error() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "cancel_cancelled_match"),
-        &Platform::Lichess,
-    );
-
-    // Cancel the match first
-    client.cancel_match(&id, &player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-
-    // Try to cancel the already cancelled match
-    let result = client.try_cancel_match(&id, &player1);
-    assert!(
-        matches!(result, Err(Ok(Error::MatchAlreadyActive)) | Err(Ok(Error::InvalidState))),
-        "Expected MatchAlreadyActive or InvalidState error when cancelling an already cancelled match"
-    );
-}
-
-#[test]
-fn test_ttl_extended_on_create_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_game1"),
-        &Platform::Lichess,
-    );
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_ttl_extended_on_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_game2"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_active_matches_ttl_refreshed_on_append_and_removal() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let match1 = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_active_append_remove_1"),
-        &Platform::Lichess,
-    );
-
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    let _match2 = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_active_append_remove_2"),
-        &Platform::Lichess,
-    );
-
-    let ttl_after_append = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::ActiveMatches)
-    });
-    assert_eq!(ttl_after_append, crate::MATCH_TTL_LEDGERS);
-
-    client.deposit(&match1, &player1);
-    client.deposit(&match1, &player2);
-
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    client.submit_result(&match1, &Winner::Player1);
-
-    let ttl_after_removal = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::ActiveMatches)
-    });
-    assert_eq!(ttl_after_removal, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_ttl_extended_on_submit_result() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_game3"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    client.submit_result(&id, &Winner::Player2);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_non_oracle_unauthorized_even_when_paused() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "paused_unauth"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    client.pause();
-
-    // A random address that is not the oracle attempts to submit a result
-    // while the contract is paused — must get Unauthorized, not ContractPaused.
-    let non_oracle = Address::generate(&env);
-    env.mock_auths(&[soroban_sdk::testutils::MockAuth {
-        address: &non_oracle,
-        invoke: &soroban_sdk::testutils::MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "submit_result",
-            args: (id, Winner::Player1).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-    let result = client.try_submit_result(&id, &Winner::Player1);
-    assert!(
-        matches!(
-            result,
-            Err(Err(_)) | Err(Ok(Error::Unauthorized)) | Err(Ok(Error::ContractPaused))
-        ),
-        "expected auth failure (Abort, Unauthorized, or ContractPaused) for non-oracle caller on paused contract"
-    );
-}
-
-#[test]
-fn test_ttl_extended_on_cancel() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_game4"),
-        &Platform::Lichess,
-    );
-
-    // Advance ledger so TTL decreases, making the subsequent extend_ttl in
-    // cancel_match meaningful — without it the assertion would pass trivially
-    // because create_match already set TTL to MATCH_TTL_LEDGERS.
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    client.cancel_match(&id, &player1);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_is_funded_extends_ttl() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_is_funded"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    // Advance ledgers so TTL would have decreased without extend
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    client.is_funded(&id);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-// #287 — created_ledger is populated on create_match
-#[test]
-fn test_created_ledger_is_set() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // Advance the ledger so sequence is non-zero
-    env.ledger().set_sequence_number(42);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ledger_game"),
-        &Platform::Lichess,
-    );
-
-    let m = client.get_match(&id);
-    assert_eq!(
-        m.created_ledger, 42,
-        "created_ledger should match ledger sequence at creation"
-    );
-}
-
-// #292 — MatchCount increments correctly across multiple matches
-#[test]
-fn test_match_count_increments_sequentially() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let game_ids = ["seq0", "seq1", "seq2", "seq3", "seq4"];
-    for (expected_id, game_id_str) in game_ids.iter().enumerate() {
-        let id = client.create_match(
-            &player1,
-            &player2,
-            &100,
-            &token,
-            &String::from_str(&env, game_id_str),
-            &Platform::Lichess,
-        );
-        assert_eq!(id, expected_id as u64);
-    }
-
-    let last = client.get_match(&4);
-    assert_eq!(last.id, 4);
-    assert_eq!(last.state, MatchState::Pending);
-}
-
-// #296 — get_escrow_balance returns 0 after draw payout
-#[test]
-fn test_escrow_balance_zero_after_draw() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "draw_balance_game"),
-        &Platform::ChessDotCom,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    assert_eq!(client.get_escrow_balance(&id), 200);
-
-    client.submit_result(&id, &Winner::Draw);
-
-    assert_eq!(client.get_escrow_balance(&id), 0);
-}
-
-#[test]
-fn test_get_escrow_balance_returns_stake_amount_after_player1_deposits() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "escrow_balance_player1"),
-        &Platform::Lichess,
-    );
-
-    // Before any deposits, escrow balance should be 0
-    assert_eq!(client.get_escrow_balance(&id), 0);
-
-    // After player1 deposits, escrow balance should be 100 (1 * stake_amount)
-    client.deposit(&id, &player1);
-    assert_eq!(client.get_escrow_balance(&id), 100);
-}
-
-#[test]
-fn test_expire_match_refunds_depositor_after_timeout() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    env.ledger().set_sequence_number(100);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "expire_game"),
-        &Platform::Lichess,
-    );
-
-    // Only player1 deposits
-    client.deposit(&id, &player1);
-
-    let p1_balance_before = token::Client::new(&env, &token).balance(&player1);
-
-    env.deployer().extend_ttl_for_contract_instance(
-        contract_id.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(contract_id.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.deployer().extend_ttl_for_contract_instance(
-        token.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(token.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    // Advance ledger past the default timeout (17_280 ledgers)
-    env.ledger().set_sequence_number(100 + 17_280);
-
-    env.deployer().extend_ttl_for_contract_instance(
-        contract_id.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(contract_id.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.deployer().extend_ttl_for_contract_instance(
-        token.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(token.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    client.expire_match(&id);
-
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Cancelled);
-
-    // player1 should have their stake back
-    let p1_balance_after = token::Client::new(&env, &token).balance(&player1);
-    assert_eq!(p1_balance_after - p1_balance_before, 100);
-}
-
-#[test]
-fn test_expire_match_fails_before_timeout() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    env.ledger().set_sequence_number(100);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "early_expire"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-
-    // Not enough ledgers have passed
-    env.ledger().set_sequence_number(100 + 100);
-
-    let result = client.try_expire_match(&id);
-    assert_eq!(result, Err(Ok(Error::MatchNotExpired)));
-}
-
-#[test]
-fn test_get_oracle_returns_initialized_address() {
-    let (env, contract_id, oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    assert_eq!(client.get_oracle(), oracle);
-}
-
-#[test]
-fn test_get_oracle_returns_updated_address_after_update_oracle() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let new_oracle = Address::generate(&env);
-    client.update_oracle(&new_oracle);
-    assert_eq!(client.get_oracle(), new_oracle);
-}
-
-#[test]
-fn test_get_match_returns_correct_players() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "players_test"),
-        &Platform::Lichess,
-    );
-
-    let m = client.get_match(&id);
-    assert_eq!(m.player1, player1);
-    assert_eq!(m.player2, player2);
-}
-
-#[test]
-fn test_get_match_timeout_returns_default() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let timeout = client.try_get_match_timeout().unwrap().unwrap();
-    assert_eq!(timeout, DEFAULT_MATCH_TIMEOUT_LEDGERS);
-}
-
-#[test]
-fn test_get_match_returns_match_not_found_for_unknown_id() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_get_match(&9999u64);
-    assert_eq!(result, Err(Ok(Error::MatchNotFound)));
-}
-
-#[test]
-fn test_update_oracle_emits_oracle_up_event_with_addresses() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let new_oracle = Address::generate(&env);
-    let old_oracle: Address = client.get_oracle();
-
-    client.update_oracle(&new_oracle);
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "admin").into_val(&env),
-        soroban_sdk::symbol_short!("oracle_up").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "oracle_up event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let (ev_old, ev_new): (Address, Address) = TryFromVal::try_from_val(&env, &data).unwrap();
-    assert_eq!(ev_old, old_oracle);
-    assert_eq!(ev_new, new_oracle);
-}
-
-#[test]
-fn test_is_funded_returns_false_when_only_player1_deposited() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "funded_test"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    assert!(!client.is_funded(&id));
-
-    client.deposit(&id, &player2);
-    assert!(client.is_funded(&id));
-}
-
-#[test]
-fn test_submit_result_on_nonexistent_match_id_returns_match_not_found() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_submit_result(&9999u64, &Winner::Player1);
-    assert_eq!(result, Err(Ok(Error::MatchNotFound)));
-}
-
-#[test]
-fn test_cancel_match_by_player2_refunds_player1_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = TokenClient::new(&env, &token);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "cancel_test"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    let player1_balance_after_deposit = token_client.balance(&player1);
-    assert_eq!(player1_balance_after_deposit, 900);
-
-    client.cancel_match(&id, &player2);
-
-    let player1_balance_after_cancel = token_client.balance(&player1);
-    assert_eq!(player1_balance_after_cancel, 1000);
-    assert_eq!(token_client.balance(&player2), 1000);
-}
-
-#[test]
-fn test_cancel_match_by_unauthorized_address_returns_unauthorized() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let third_party = Address::generate(&env);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "unauthorized_cancel_test"),
-        &Platform::Lichess,
-    );
-
-    let result = client.try_cancel_match(&id, &third_party);
-    assert_eq!(result, Err(Ok(Error::Unauthorized)));
-}
-
-// #373 — update_oracle routes subsequent submit_result to the new oracle
-#[test]
-fn test_update_oracle_routes_submit_result() {
-    let (env, contract_id, oracle_old, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let oracle_new = Address::generate(&env);
-    client.update_oracle(&oracle_new);
-    assert_eq!(client.get_oracle(), oracle_new);
-
-    // Match for oracle_new success assertion
-    let id1 = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "oracle_new_match"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id1, &player1);
-    client.deposit(&id1, &player2);
-
-    // oracle_new must succeed
-    env.mock_auths(&[MockAuth {
-        address: &oracle_new,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "submit_result",
-            args: (id1, Winner::Player1).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-    client.submit_result(&id1, &Winner::Player1);
-    assert_eq!(client.get_match(&id1).state, MatchState::Completed);
-
-    // Re-enable all auths for token minting
-    env.mock_all_auths();
-
-    // Match for oracle_old rejection assertion
-    let asset_client = StellarAssetClient::new(&env, &token);
-    asset_client.mint(&player1, &100);
-    asset_client.mint(&player2, &100);
-    let id2 = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "oracle_old_match"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id2, &player1);
-    client.deposit(&id2, &player2);
-
-    // oracle_old must be rejected
-    env.mock_auths(&[MockAuth {
-        address: &oracle_old,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "submit_result",
-            args: (id2, Winner::Player1).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-    let result = client.try_submit_result(&id2, &Winner::Player1);
-    assert!(
-        matches!(result, Err(Err(_))),
-        "old oracle must be rejected after rotation"
-    );
-}
-
-#[test]
-fn test_submit_result_from_non_oracle_returns_unauthorized() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "non_oracle_submit_game"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    let non_oracle = Address::generate(&env);
-    env.mock_auths(&[MockAuth {
-        address: &non_oracle,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "submit_result",
-            args: (id, Winner::Player1).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-
-    let result = client.try_submit_result(&id, &Winner::Player1);
-    assert!(
-        matches!(result, Err(Err(_)) | Err(Ok(Error::Unauthorized))),
-        "expected auth failure for non-oracle caller"
-    );
-}
-
-#[test]
-fn test_get_match_returns_winner_after_payout() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "winner_test"),
-        &Platform::Lichess,
-    );
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    client.submit_result(&id, &Winner::Player2);
-
-    let m = client.get_match(&id);
-    assert_eq!(m.winner, Winner::Player2);
-}
-
-#[test]
-fn test_submit_result_overflow_on_extreme_stake() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // Create a match with a normal stake, then directly overwrite the stake_amount
-    // in storage to i128::MAX so that stake_amount * 2 overflows — bypassing the
-    // token layer which would also overflow on deposit.
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "overflow_game"),
-        &Platform::Lichess,
-    );
-
-    env.as_contract(&contract_id, || {
-        let mut m: Match = env.storage().persistent().get(&DataKey::Match(id)).unwrap();
-        m.stake_amount = i128::MAX;
-        m.state = MatchState::Active;
-        m.player1_deposited = true;
-        m.player2_deposited = true;
-        env.storage().persistent().set(&DataKey::Match(id), &m);
-    });
-
-    let result = client.try_submit_result(&id, &Winner::Player1);
-    assert_eq!(result, Err(Ok(Error::Overflow)));
-}
-
-#[test]
-fn test_two_step_admin_transfer() {
-    let (env, contract_id, _oracle, _p1, _p2, _token, admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let new_admin = Address::generate(&env);
-
-    // Step 1: propose — old admin is still active
-    client.propose_admin(&new_admin);
-    assert_eq!(client.get_admin(), admin);
-
-    // Step 2: accept — new admin takes over
-    client.accept_admin();
-    assert_eq!(client.get_admin(), new_admin);
-
-    // Old admin is now rejected — clear mocks so auth is enforced
-    env.set_auths(&[]);
-    let result = client.try_propose_admin(&admin);
-    assert!(result.is_err());
-}
-
-#[test]
-fn test_transfer_admin_pause_auth() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let new_admin = Address::generate(&env);
-
-    // Transfer admin to new_admin (mock_all_auths is active from setup)
-    client.transfer_admin(&new_admin);
-    assert_eq!(client.get_admin(), new_admin);
-
-    // Old admin tries to pause — must fail
-    env.mock_auths(&[MockAuth {
-        address: &admin,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "pause",
-            args: ().into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-    let result = client.try_pause();
-    assert!(
-        result.is_err(),
-        "old admin should be rejected from pause after transfer"
-    );
-
-    // New admin pauses — must succeed
-    env.mock_auths(&[MockAuth {
-        address: &new_admin,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "pause",
-            args: ().into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-    client.pause();
-}
-
-// ── Issue: is_funded checks deposit flags not state ─────────────────────────
-//
-// `is_funded` inspects `player1_deposited && player2_deposited` rather than
-// the match state. This means it returns `true` even after payout, because the
-// deposit flags are never cleared when the match transitions to Completed.
-//
-// Expected behaviour (documented here):
-//   - Before both deposits  → false
-//   - After both deposits   → true  (match is Active)
-//   - After payout          → true  (flags still set; state is Completed)
-//
-// Callers that need to know whether funds are *currently held in escrow* should
-// use `get_escrow_balance` (returns 0 for Completed/Cancelled) or check
-// `get_match().state == MatchState::Active` instead of relying on `is_funded`.
-#[test]
-fn test_is_funded_returns_true_after_payout() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "is_funded_post_payout"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    // Both deposited — match is Active, is_funded must be true
-    assert!(
-        client.is_funded(&id),
-        "is_funded must be true when both players have deposited"
-    );
-    assert_eq!(client.get_match(&id).state, MatchState::Active);
-
-    // Complete the match
-    client.submit_result(&id, &Winner::Player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Completed);
-
-    // is_funded still returns true because it checks deposit flags, not state.
-    // This is the documented (if surprising) behaviour: the flags are never
-    // cleared on payout. Use get_escrow_balance or check state directly when
-    // you need to know whether funds are still held.
-    assert!(
-        client.is_funded(&id),
-        "is_funded returns true after payout because it checks deposit flags, not match state"
-    );
-
-    // Confirm that get_escrow_balance correctly returns 0 for a Completed match
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        0,
-        "get_escrow_balance must return 0 for a Completed match"
-    );
-}
-
-// ── Issue: get_escrow_balance returns 0 for Completed matches ────────────────
-//
-// No test previously verified this specifically. The implementation short-circuits
-// to 0 for Completed and Cancelled states, but that path was untested.
-#[test]
-fn test_get_escrow_balance_zero_for_completed_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "balance_completed"),
-        &Platform::Lichess,
-    );
-
-    // Fund both players
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        200,
-        "escrow balance must be 2x stake while Active"
-    );
-
-    // Complete the match — payout transfers funds out
-    client.submit_result(&id, &Winner::Player2);
-    assert_eq!(client.get_match(&id).state, MatchState::Completed);
-
-    // get_escrow_balance must return 0 for a Completed match
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        0,
-        "get_escrow_balance must return 0 after match is Completed"
-    );
-}
-
-// ── Issue: get_escrow_balance returns 0 for Cancelled match with no deposits ─
-//
-// No test previously verified the case where a match is cancelled before any
-// deposits are made. The balance should be 0 both because no funds were ever
-// transferred in and because the Cancelled state short-circuits to 0.
-#[test]
-fn test_get_escrow_balance_zero_for_cancelled_match_no_deposits() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "balance_cancelled_no_deposit"),
-        &Platform::Lichess,
-    );
-
-    // No deposits made — cancel immediately
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        0,
-        "escrow balance must be 0 before any deposits"
-    );
-    client.cancel_match(&id, &player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-
-    // get_escrow_balance must return 0 for a Cancelled match with no deposits
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        0,
-        "get_escrow_balance must return 0 for a Cancelled match where no deposits were made"
-    );
-}
-
-#[test]
-fn test_get_escrow_balance_zero_after_cancel_with_player1_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "balance_cancelled_after_player1_deposit"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        100,
-        "escrow balance must reflect player1's deposited stake before cancellation"
-    );
-
-    client.cancel_match(&id, &player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        0,
-        "get_escrow_balance must return 0 after cancelling a match and refunding player1"
-    );
-}
-
-#[test]
-fn test_ttl_extended_on_get_escrow_balance() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_balance_game"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-
-    // Get the TTL before calling get_escrow_balance
-    let ttl_before = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-
-    // Call get_escrow_balance which should extend TTL
-    let _balance = client.get_escrow_balance(&id);
-
-    // Get the TTL after calling get_escrow_balance
-    let ttl_after = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-
-    // TTL should be extended (increased)
-    assert!(
-        ttl_after >= ttl_before,
-        "TTL should be extended after get_escrow_balance"
-    );
-}
-
-#[test]
-fn test_deposit_after_cancel_match_returns_invalid_state() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "deposit_after_cancel"),
-        &Platform::Lichess,
-    );
-
-    client.cancel_match(&id, &player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Cancelled);
-
-    let result = client.try_deposit(&id, &player2);
-    assert_eq!(result, Err(Ok(Error::InvalidState)));
-}
-
-#[test]
-fn test_match_state_active_after_both_deposits() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "active_state_test"),
-        &Platform::Lichess,
-    );
-
-    // Initially pending
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Pending);
-
-    // After player1 deposits, still pending
-    client.deposit(&id, &player1);
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Pending);
-
-    // After both deposits, becomes active
-    client.deposit(&id, &player2);
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Active);
-}
-
-#[test]
-fn test_create_match_rejects_same_player_as_both_sides() {
-    let (env, contract_id, _oracle, player1, _player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_create_match(
-        &player1,
-        &player1,
-        &100,
-        &token,
-        &String::from_str(&env, "self_match"),
-        &Platform::Lichess,
-    );
-    assert_eq!(result, Err(Ok(Error::InvalidPlayers)));
-}
-
-/// get_match extends TTL on every read so hot matches never expire between reads.
-#[test]
-fn test_get_match_extends_ttl_on_read() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_read_test"),
-        &Platform::Lichess,
-    );
-
-    client.get_match(&id);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, MATCH_TTL_LEDGERS);
-}
-
-/// get_match resets TTL to MATCH_TTL_LEDGERS even after ledgers have advanced.
-#[test]
-fn test_get_match_resets_ttl_after_ledger_advance() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_get_match"),
-        &Platform::Lichess,
-    );
-
-    // Advance ledger by 1000 so the TTL has decreased
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    client.get_match(&id);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage().persistent().get_ttl(&DataKey::Match(id))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_get_match_returns_cancelled_after_expire_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    env.ledger().set_sequence_number(100);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "expire_state_game"),
-        &Platform::Lichess,
-    );
-
-    // Extend TTLs so storage survives the ledger jump
-    for addr in [&contract_id, &token] {
-        env.deployer().extend_ttl_for_contract_instance(
-            addr.clone(),
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-        env.deployer()
-            .extend_ttl_for_code(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    }
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    // Advance past the 17_280-ledger timeout
-    env.ledger().set_sequence_number(100 + 17_280);
-
-    for addr in [&contract_id, &token] {
-        env.deployer().extend_ttl_for_contract_instance(
-            addr.clone(),
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-        env.deployer()
-            .extend_ttl_for_code(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    }
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    client.expire_match(&id);
-
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Cancelled);
-}
-
-#[test]
-fn test_update_oracle_rejects_non_admin_caller() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let attacker = Address::generate(&env);
-    let new_oracle = Address::generate(&env);
-
-    env.mock_auths(&[MockAuth {
-        address: &attacker,
-        invoke: &MockAuthInvoke {
-            contract: &contract_id,
-            fn_name: "update_oracle",
-            args: (new_oracle.clone(),).into_val(&env),
-            sub_invokes: &[],
-        },
-    }]);
-
-    let result = client.try_update_oracle(&new_oracle);
-    assert!(
-        result.is_err(),
-        "update_oracle must reject a non-admin caller"
-    );
-}
-
-#[test]
-fn test_submit_result_emits_completed_event_with_correct_winner() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let match_id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "event_test"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&match_id, &player1);
-    client.deposit(&match_id, &player2);
-
-    client.submit_result(&match_id, &Winner::Player1);
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "match").into_val(&env),
-        symbol_short!("completed").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "match completed event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let (ev_match_id, ev_winner): (u64, Winner) = TryFromVal::try_from_val(&env, &data).unwrap();
-    assert_eq!(ev_match_id, match_id);
-    assert_eq!(ev_winner, Winner::Player1);
-}
-
-#[test]
-fn test_create_match_with_chess_dot_com_platform() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "chess_dot_com_game"),
-        &Platform::ChessDotCom,
-    );
-
-    let m = client.get_match(&id);
-    assert_eq!(m.platform, Platform::ChessDotCom);
-}
-
-#[test]
-fn test_double_deposit() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "double_deposit_test"),
-        &Platform::Lichess,
-    );
-
-    // First deposit should succeed
-    client.deposit(&id, &player1);
-    assert!(!client.is_funded(&id));
-
-    // Second deposit by player1 should fail with AlreadyFunded
-    let result = client.try_deposit(&id, &player1);
-    assert_eq!(result, Err(Ok(Error::AlreadyFunded)));
-}
-
-#[test]
-fn test_is_paused_cycle() {
-    let (env, contract_id, _oracle, _player1, _player2, _token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    assert!(!client.is_paused());
-    client.pause();
-    assert!(client.is_paused());
-    client.unpause();
-    assert!(!client.is_paused());
-}
-
-#[test]
-fn test_initialize_rejects_self_as_oracle() {
-    let env = Env::default();
-    env.mock_all_auths();
-
-    let admin = Address::generate(&env);
-    let contract_id = env.register_contract(None, EscrowContract);
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let result = client.try_initialize(&contract_id, &admin);
-    assert_eq!(result, Err(Ok(Error::InvalidAddress)));
-}
-
-/// PlayerMatches index TTL is refreshed when a new match is appended (write path).
-#[test]
-fn test_player_match_index_ttl_refreshes_on_append() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_index_append_1"),
-        &Platform::Lichess,
-    );
-
-    // Advance ledger so TTL would have decayed without an extend
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    // Appending a second match must refresh the index TTL
-    client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_index_append_2"),
-        &Platform::Lichess,
-    );
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage()
-            .persistent()
-            .get_ttl(&DataKey::PlayerMatches(player1.clone()))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-/// PlayerMatches index TTL is refreshed when the index is read (getter path).
-#[test]
-fn test_player_match_index_ttl_refreshes_on_read() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "ttl_index_read"),
-        &Platform::Lichess,
-    );
-
-    // Advance ledger so TTL would have decayed without an extend
-    env.ledger().set(soroban_sdk::testutils::LedgerInfo {
-        sequence_number: env.ledger().sequence() + 1000,
-        timestamp: env.ledger().timestamp() + 5000,
-        protocol_version: 22,
-        network_id: Default::default(),
-        base_reserve: 10,
-        min_temp_entry_ttl: 1,
-        min_persistent_entry_ttl: 1,
-        max_entry_ttl: crate::MATCH_TTL_LEDGERS + 2000,
-    });
-
-    client.get_player_matches(&player1);
-
-    let ttl = env.as_contract(&contract_id, || {
-        env.storage()
-            .persistent()
-            .get_ttl(&DataKey::PlayerMatches(player1.clone()))
-    });
-    assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-}
-
-#[test]
-fn test_expire_match_refunds_both_players_when_both_deposited_but_still_pending() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = token::Client::new(&env, &token);
-
-    env.ledger().set_sequence_number(100);
-
-    // Both players deposit, but we manually keep the state Pending to simulate
-    // the scenario where both deposited yet the match never transitioned to Active.
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "expire_both_deposited"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id, &player1);
-    client.deposit(&id, &player2);
-
-    // At this point the contract transitions to Active after both deposits.
-    // Force the state back to Pending to represent the target scenario.
-    env.as_contract(&contract_id, || {
-        let mut m: Match = env.storage().persistent().get(&DataKey::Match(id)).unwrap();
-        m.state = MatchState::Pending;
-        env.storage().persistent().set(&DataKey::Match(id), &m);
-    });
-
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Pending);
-    assert!(m.player1_deposited);
-    assert!(m.player2_deposited);
-
-    let p1_balance_before = token_client.balance(&player1);
-    let p2_balance_before = token_client.balance(&player2);
-
-    // Extend TTLs so storage entries survive the ledger jump
-    env.deployer().extend_ttl_for_contract_instance(
-        contract_id.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(contract_id.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.deployer().extend_ttl_for_contract_instance(
-        token.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(token.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    // Advance ledger past the default timeout (17_280 ledgers)
-    env.ledger().set_sequence_number(100 + 17_280);
-
-    env.deployer().extend_ttl_for_contract_instance(
-        contract_id.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(contract_id.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.deployer().extend_ttl_for_contract_instance(
-        token.clone(),
-        MATCH_TTL_LEDGERS,
-        MATCH_TTL_LEDGERS,
-    );
-    env.deployer()
-        .extend_ttl_for_code(token.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    client.expire_match(&id);
-
-    // Match must be Cancelled
-    let m = client.get_match(&id);
-    assert_eq!(m.state, MatchState::Cancelled);
-
-    // Both players must be fully refunded
-    assert_eq!(token_client.balance(&player1) - p1_balance_before, 100);
-    assert_eq!(token_client.balance(&player2) - p2_balance_before, 100);
-}
-
-
-// ── Task #1: expire_match emits ("match", "expired") with match_id payload ──
-#[test]
-fn test_expire_match_emits_expired_event() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    env.ledger().set_sequence_number(100);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-
-        &String::from_str(&env, "is_funded_completion"),
-        &Platform::Lichess,
-    );
-
-    assert!(!client.is_funded(&id), "unfunded before any deposit");
-
-    client.deposit(&id, &player1);
-    assert!(!client.is_funded(&id), "still unfunded after only player1 deposits");
-
-    client.deposit(&id, &player2);
-    assert!(client.is_funded(&id), "funded after both players deposit");
-    assert_eq!(client.get_match(&id).state, MatchState::Active);
-
-    client.submit_result(&id, &Winner::Player1);
-    assert_eq!(client.get_match(&id).state, MatchState::Completed);
-
-    // Funds are no longer held — escrow balance is the authoritative check.
-    assert_eq!(
-        client.get_escrow_balance(&id),
-        0,
-        "escrow balance must be 0 after completion"
-    );
-    // is_funded still reflects deposit flags (documented behaviour); callers
-    // needing "funds currently in escrow" must use get_escrow_balance or state.
-    assert!(
-        !client.is_funded(&id) || client.get_escrow_balance(&id) == 0,
-        "after completion, either is_funded is false or escrow balance is 0"
-    );
-}
-
-// ── Issue #15: default winner is Draw (Undecided) immediately after creation ──
-//
-// The Match struct initialises `winner` to `Winner::Draw` as a stand-in for
-// "no result yet". This test asserts that a freshly created match carries that
-// default before any result is submitted.
-#[test]
-fn test_winner_is_draw_default_before_result_submitted() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    let id = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "default_winner_test"),
-        &Platform::Lichess,
-    );
-
-    let m = client.get_match(&id);
-    assert_eq!(
-        m.state,
-        MatchState::Pending,
-        "match must be Pending immediately after creation"
-    );
-    assert_eq!(
-        m.winner,
-        Winner::Draw,
-        "winner must default to Draw (Undecided) before any result is submitted"
-    );
-}
-
-// ── Issues #10 / #11: terminal paths — cancel and expire, state + ledger metadata ──
-//
-// Covers both terminal paths in a single test group:
-//   • cancel_match  → state == Cancelled, completed_ledger is None
-//   • expire_match  → state == Cancelled, completed_ledger is None
-//
-// Neither path sets completed_ledger (only submit_result does), so both must
-// leave that field as None.
-#[test]
-fn test_cancel_and_expire_terminal_state_and_ledger_metadata() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-    let token_client = token::Client::new(&env, &token);
-
-    // ── cancel path ──────────────────────────────────────────────────────────
-    let id_cancel = client.create_match(
-        &player1,
-        &player2,
-        &100,
-        &token,
-        &String::from_str(&env, "terminal_cancel"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id_cancel, &player1);
-    client.cancel_match(&id_cancel, &player1);
-
-    let m_cancel = client.get_match(&id_cancel);
-    assert_eq!(m_cancel.state, MatchState::Cancelled, "cancel → Cancelled");
-    assert!(
-        m_cancel.completed_ledger.is_none(),
-        "cancel must not set completed_ledger"
-    );
-    // Refund verified
-    assert_eq!(token_client.balance(&player1), 1000);
-
-    // ── expire path ──────────────────────────────────────────────────────────
-    env.ledger().set_sequence_number(500);
-
-    let id_expire = client.create_match(
-
-        &String::from_str(&env, "expire_event_game"),
-        &Platform::Lichess,
-    );
-
-    // Extend TTLs so storage survives the ledger jump
-    for addr in [&contract_id, &token] {
-        env.deployer()
-            .extend_ttl_for_contract_instance(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-        env.deployer()
-            .extend_ttl_for_code(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    }
-    env.as_contract(&contract_id, || {
-        env.storage()
-            .persistent()
-            .extend_ttl(&DataKey::ActiveMatches, MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    });
-
-    // Advance past the default timeout
-    env.ledger().set_sequence_number(100 + DEFAULT_MATCH_TIMEOUT_LEDGERS);
-
-    for addr in [&contract_id, &token] {
-        env.deployer()
-            .extend_ttl_for_contract_instance(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-        env.deployer()
-            .extend_ttl_for_code(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    }
-    env.as_contract(&contract_id, || {
-        env.storage()
-            .persistent()
-            .extend_ttl(&DataKey::ActiveMatches, MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    });
-
-    client.expire_match(&id);
-
-    let events = env.events().all();
-    let expected_topics = vec![
-        &env,
-        Symbol::new(&env, "match").into_val(&env),
-        symbol_short!("expired").into_val(&env),
-    ];
-    let matched = events
-        .iter()
-        .find(|(_, topics, _)| *topics == expected_topics);
-    assert!(matched.is_some(), "match expired event not emitted");
-
-    let (_, _, data) = matched.unwrap();
-    let ev_id: u64 = TryFromVal::try_from_val(&env, &data).unwrap();
-    assert_eq!(ev_id, id);
-}
-
-// ── Task #2: lowering timeout after match creation affects expiry immediately ─
-#[test]
-fn test_lowering_timeout_after_match_creation_affects_expiry_immediately() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    env.ledger().set_sequence_number(100);
-
-    let id = client.create_match(
-// #618 — instance TTL is refreshed after create_match
-#[test]
-fn test_instance_ttl_refreshed_after_create_match() {
-    let (env, contract_id, _oracle, player1, player2, token, _admin) = setup();
-    let client = EscrowContractClient::new(&env, &contract_id);
-
-    // Advance the ledger so the instance TTL decreases from its post-initialize value,
-    // giving us a meaningful "before" baseline that is strictly less than MATCH_TTL_LEDGERS.
-    env.ledger().with_mut(|l| l.sequence_number += 1000);
-
-    let ttl_before = env.as_contract(&contract_id, || {
-        env.storage().instance().get_ttl()
-    });
-
-    client.create_match(
-
-        &player1,
-        &player2,
-        &100,
-        &token,
-
-        &String::from_str(&env, "terminal_expire"),
-        &Platform::Lichess,
-    );
-
-    client.deposit(&id_expire, &player1);
-
-    let p1_before = token_client.balance(&player1);
-
-    // Extend TTLs so storage survives the ledger jump
-    for addr in [&contract_id, &token] {
-        env.deployer().extend_ttl_for_contract_instance(
-            addr.clone(),
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-        env.deployer()
-            .extend_ttl_for_code(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    }
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    env.ledger().set_sequence_number(500 + DEFAULT_MATCH_TIMEOUT_LEDGERS);
-
-    for addr in [&contract_id, &token] {
-        env.deployer().extend_ttl_for_contract_instance(
-            addr.clone(),
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-        env.deployer()
-            .extend_ttl_for_code(addr.clone(), MATCH_TTL_LEDGERS, MATCH_TTL_LEDGERS);
-    }
-    env.as_contract(&contract_id, || {
-        env.storage().persistent().extend_ttl(
-            &DataKey::ActiveMatches,
-            MATCH_TTL_LEDGERS,
-            MATCH_TTL_LEDGERS,
-        );
-    });
-
-    client.expire_match(&id_expire);
-
-    let m_expire = client.get_match(&id_expire);
-    assert_eq!(m_expire.state, MatchState::Cancelled, "expire → Cancelled");
-    assert!(
-        m_expire.completed_ledger.is_none(),
-        "expire must not set completed_ledger"
-    );
-    // Refund verified
-    assert_eq!(token_client.balance(&player1) - p1_before, 100);
-}
-
-        &String::from_str(&env, "instance_ttl_game"),
-        &Platform::Lichess,
-    );
-
-    let ttl_after = env.as_contract(&contract_id, || {
-        env.storage().instance().get_ttl()
-    });
-
-    assert!(
-        ttl_after > ttl_before,
-        "instance TTL should be strictly greater after create_match (was {ttl_before}, now {ttl_after})"
-    );
-    assert_eq!(
-        ttl_after,
-        crate::MATCH_TTL_LEDGERS,
-        "instance TTL should be extended to MATCH_TTL_LEDGERS"
-    );
-}
-
-
