+++ contracts/escrow/src/tests/mod.rs
@@ -1,0 +1,47 @@
+extern crate std;
+
+pub use super::*;
+pub use soroban_sdk::{
+    testutils::{MockAuth, MockAuthInvoke},
+    token::{Client as TokenClient, StellarAssetClient},
+    vec, Address, Env, IntoVal, String, Symbol, TryFromVal,
+};
+
+mod admin;
+mod events;
+mod lifecycle;
+mod ttl;
+
+pub fn setup() -> (Env, Address, Address, Address, Address, Address, Address) {
+    let env = Env::default();
+    env.mock_all_auths();
+
+    let admin = Address::generate(&env);
+    let oracle = Address::generate(&env);
+    let player1 = Address::generate(&env);
+    let player2 = Address::generate(&env);
+
+    let token_id = env.register_stellar_asset_contract_v2(admin.clone());
+    let token_addr = token_id.address();
+    let asset_client = StellarAssetClient::new(&env, &token_addr);
+    asset_client.mint(&player1, &1000);
+    asset_client.mint(&player2, &1000);
+
+    let contract_id = env.register_contract(None, EscrowContract);
+    let client = EscrowContractClient::new(&env, &contract_id);
+    client.initialize(&oracle, &admin);
+
+    (
+        env,
+        contract_id,
+        oracle,
+        player1,
+        player2,
+        token_addr,
+        admin,
+    )
+}
+
+pub fn mint_player_balance(asset_client: &StellarAssetClient, player: &Address, amount: i128) {
+    asset_client.mint(player, &amount);
+}
