+++ contracts/oracle/src/lib.rs
@@ -247,726 +247,5 @@ impl OracleContract {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use escrow::types::{MatchState, Platform, Winner as EscrowWinner};
-    use escrow::{EscrowContract, EscrowContractClient};
-    use soroban_sdk::{
-        testutils::storage::{Instance as _, Persistent as _},
-        testutils::{Address as _, Events as _},
-        token::StellarAssetClient,
-        Address, Env, IntoVal, String, Symbol,
-    };
+mod tests;
 
-    fn setup() -> (Env, Address, Address, Address, Address, Address, Address) {
-        let env = Env::default();
-        env.mock_all_auths();
-
-        let admin = Address::generate(&env);
-        let oracle_admin = Address::generate(&env);
-        let player1 = Address::generate(&env);
-        let player2 = Address::generate(&env);
-
-        // Register token
-        let token_id = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_addr = token_id.address();
-        let asset_client = StellarAssetClient::new(&env, &token_addr);
-        asset_client.mint(&player1, &1000);
-        asset_client.mint(&player2, &1000);
-
-        // Register escrow contract and create + fund a match (id=0)
-        let escrow_id = env.register_contract(None, EscrowContract);
-        let escrow_client = EscrowContractClient::new(&env, &escrow_id);
-        escrow_client.initialize(&oracle_admin, &admin);
-        escrow_client.create_match(
-            &player1,
-            &player2,
-            &100,
-            &token_addr,
-            &String::from_str(&env, "test_game"),
-            &Platform::Lichess,
-        );
-        escrow_client.deposit(&0u64, &player1);
-        escrow_client.deposit(&0u64, &player2);
-
-        // Register oracle contract
-        let oracle_id = env.register_contract(None, OracleContract);
-        let oracle_client = OracleContractClient::new(&env, &oracle_id);
-        oracle_client.initialize(&oracle_admin);
-
-        (
-            env,
-            oracle_id,
-            escrow_id,
-            oracle_admin,
-            player1,
-            player2,
-            token_addr,
-        )
-    }
-
-    #[test]
-    fn test_initialize_emits_event() {
-        let env = Env::default();
-        env.mock_all_auths();
-
-        let admin = Address::generate(&env);
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-        client.initialize(&admin);
-
-        let events = env.events().all();
-        let expected_topics = soroban_sdk::vec![
-            &env,
-            Symbol::new(&env, "oracle").into_val(&env),
-            symbol_short!("init").into_val(&env),
-        ];
-        let matched = events
-            .iter()
-            .find(|(_, topics, _)| *topics == expected_topics);
-        assert!(matched.is_some(), "oracle initialized event not emitted");
-
-        let (_, _, data) = matched.unwrap();
-        let ev_admin: Address = soroban_sdk::TryFromVal::try_from_val(&env, &data).unwrap();
-        assert_eq!(ev_admin, admin);
-    }
-
-    // ── has_result (public, unauthenticated) ─────────────────────────────────
-
-    /// Test that has_result returns false for match_id 0 on a fresh contract.
-    #[test]
-    fn test_has_result_returns_false_for_match_id_0_on_fresh_contract() {
-        let (env, contract_id, _escrow_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // On a fresh contract, has_result(0) should return false
-        assert!(!client.has_result(&0u64));
-    }
-
-    /// Confirms that any caller can invoke has_result without authentication.
-    /// Returns false before a result is submitted and true afterwards.
-    #[test]
-    fn test_has_result_is_public_and_unauthenticated() {
-        let (env, contract_id, _escrow_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Before submission — any caller can probe, no auth required
-        assert!(!client.has_result(&0u64));
-        assert!(!client.has_result(&999u64));
-
-        client.submit_result(
-            &0u64,
-            &String::from_str(&env, "test_game"),
-            &Winner::Player1,
-        );
-
-        // After submission — still public, now returns true
-        assert!(client.has_result(&0u64));
-        // Unrelated match_id still false
-        assert!(!client.has_result(&999u64));
-    }
-
-    // ── has_result_admin (admin-gated) ────────────────────────────────────────
-
-    /// Admin can probe result existence via the gated variant.
-    #[test]
-    fn test_has_result_admin_returns_false_before_submission() {
-        let (env, contract_id, _escrow_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        assert!(!client.has_result_admin(&0u64));
-        assert!(!client.has_result_admin(&999u64));
-    }
-
-    /// has_result_admin returns true after a result is submitted.
-    #[test]
-    fn test_has_result_admin_returns_true_after_submission() {
-        let (env, contract_id, _escrow_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(
-            &0u64,
-            &String::from_str(&env, "test_game"),
-            &Winner::Player1,
-        );
-
-        assert!(client.has_result_admin(&0u64));
-    }
-
-    /// Non-admin callers must not be able to call has_result_admin.
-    #[test]
-    #[should_panic]
-    fn test_has_result_admin_rejects_non_admin() {
-        let env = Env::default();
-        // Do NOT mock all auths — we want auth to actually be enforced
-        let admin = Address::generate(&env);
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-        client.initialize(&admin);
-        // Non-admin tries to call has_result_admin - should panic
-        client.has_result_admin(&0u64);
-    }
-
-    #[test]
-    fn test_submit_and_get_result() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-
-        assert!(client.has_result(&0u64));
-        let entry = client.get_result(&0u64);
-        assert_eq!(entry.result, Winner::Player1);
-    }
-
-    /// Issue #563 — ResultEntry must record the ledger at which the result was submitted.
-    #[test]
-    fn test_submit_result_stores_submitted_ledger() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        let ledger_before = env.ledger().sequence();
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-
-        let entry = client.get_result(&0u64);
-        assert!(
-            entry.submitted_ledger >= ledger_before,
-            "submitted_ledger must be >= ledger at call time"
-        );
-    }
-
-    /// Issue #564 — ResultEntry must record the admin address that submitted the result.
-    #[test]
-    fn test_submit_result_stores_submitter() {
-        let (env, contract_id, _escrow_id, oracle_admin, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-
-        let entry = client.get_result(&0u64);
-        assert_eq!(entry.submitter, oracle_admin);
-    }
-
-    #[test]
-    fn test_submit_result_emits_event() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-
-        let events = env.events().all();
-        let expected_topics = soroban_sdk::vec![
-            &env,
-            Symbol::new(&env, "oracle").into_val(&env),
-            symbol_short!("result").into_val(&env),
-        ];
-        let matched = events
-            .iter()
-            .find(|(_, topics, _)| *topics == expected_topics);
-        assert!(matched.is_some(), "oracle result event not emitted");
-
-        let (_, _, data) = matched.unwrap();
-        let (ev_id, ev_result): (u64, Winner) =
-            soroban_sdk::TryFromVal::try_from_val(&env, &data).unwrap();
-        assert_eq!(ev_id, 0u64);
-        assert_eq!(ev_result, Winner::Player1);
-    }
-
-    #[test]
-    fn test_submit_draw_result_emits_event() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Draw);
-
-        let events = env.events().all();
-        let expected_topics = soroban_sdk::vec![
-            &env,
-            Symbol::new(&env, "oracle").into_val(&env),
-            symbol_short!("result").into_val(&env),
-        ];
-        let matched = events
-            .iter()
-            .find(|(_, topics, _)| *topics == expected_topics);
-        assert!(
-            matched.is_some(),
-            "oracle result event not emitted for Draw"
-        );
-
-        let (_, _, data) = matched.unwrap();
-        let (ev_id, ev_result): (u64, Winner) =
-            soroban_sdk::TryFromVal::try_from_val(&env, &data).unwrap();
-        assert_eq!(ev_id, 0u64);
-        assert_eq!(ev_result, Winner::Draw);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_duplicate_submit_fails() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Draw);
-        // second submit should panic
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Draw);
-    }
-
-    #[test]
-    fn test_duplicate_submit_returns_already_submitted() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Draw);
-        let result =
-            client.try_submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Draw);
-        assert_eq!(result, Err(Ok(Error::AlreadySubmitted)));
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_double_initialize_fails() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let admin = Address::generate(&env);
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.initialize(&admin);
-        // second initialize should panic
-        client.initialize(&admin);
-    }
-
-    #[test]
-    fn test_submit_result_on_uninitialized_contract_returns_unauthorized() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // No initialize call — Admin key is absent
-        let result =
-            client.try_submit_result(&0u64, &String::from_str(&env, "game_abc"), &Winner::Player1);
-        assert_eq!(result, Err(Ok(Error::Unauthorized)));
-    }
-
-    #[test]
-    fn test_is_initialized() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let admin = Address::generate(&env);
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        assert!(!client.is_initialized());
-        client.initialize(&admin);
-        assert!(client.is_initialized());
-    }
-
-    #[test]
-    fn test_ttl_extended_on_submit_result() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-
-        let ttl = env.as_contract(&contract_id, || {
-            env.storage().persistent().get_ttl(&DataKey::Result(0u64))
-        });
-        assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-    }
-
-    /// Test that get_result returns ResultNotFound for non-existent match IDs.
-    /// This verifies the invariant: querying an unknown match_id must always
-    /// return Error::ResultNotFound rather than panicking or returning invalid data.
-    #[test]
-    #[should_panic(expected = "Error(Contract, #3)")]
-    fn test_get_result_not_found() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Query a match_id that has never been submitted
-        client.get_result(&9999u64);
-    }
-
-    #[test]
-    fn test_pause_on_uninitialized_contract_returns_unauthorized() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // No initialize call — Admin key is absent
-        let result = client.try_pause();
-        assert_eq!(result, Err(Ok(Error::Unauthorized)));
-    }
-
-    /// Test that pause can only be called by admin.
-    #[test]
-    fn test_pause_admin_only() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Admin can pause
-        client.pause();
-
-        // Verify it's paused by trying to submit a result
-        let result =
-            client.try_submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-    }
-
-    /// Test that unpause can only be called by admin.
-    #[test]
-    fn test_unpause_admin_only() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Pause first
-        client.pause();
-
-        // Admin can unpause
-        client.unpause();
-
-        // Verify it's unpaused by submitting a result
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-        assert!(client.has_result(&0u64));
-    }
-
-    /// Test that submit_result returns ContractPaused when paused.
-    #[test]
-    fn test_submit_result_blocked_when_paused() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Pause the contract
-        client.pause();
-
-        // Try to submit a result - should fail with ContractPaused
-        let result =
-            client.try_submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-
-        // Verify no result was stored
-        assert!(!client.has_result(&0u64));
-    }
-
-    /// Test that submit_result works normally after unpause.
-    #[test]
-    fn test_submit_result_works_after_unpause() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Pause the contract
-        client.pause();
-
-        // Verify submit is blocked
-        let result =
-            client.try_submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-
-        // Unpause
-        client.unpause();
-
-        // Now submit should work
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-        assert!(client.has_result(&0u64));
-        let entry = client.get_result(&0u64);
-        assert_eq!(entry.result, Winner::Player1);
-    }
-
-    /// Test pause/unpause state transitions.
-    #[test]
-    fn test_pause_unpause_state_transitions() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Initially unpaused - submit should work
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-        assert!(client.has_result(&0u64));
-
-        // Pause
-        client.pause();
-
-        // Submit should fail
-        let result =
-            client.try_submit_result(&1u64, &String::from_str(&env, "def456"), &Winner::Player2);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-
-        // Unpause
-        client.unpause();
-
-        // Submit should work again
-        client.submit_result(&1u64, &String::from_str(&env, "def456"), &Winner::Player2);
-        assert!(client.has_result(&1u64));
-
-        // Can pause again
-        client.pause();
-        let result =
-            client.try_submit_result(&2u64, &String::from_str(&env, "ghi789"), &Winner::Draw);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-    }
-
-    /// Test that get_result extends TTL on read.
-    /// This prevents active results from expiring while they're still being accessed.
-    #[test]
-    fn test_get_result_extends_ttl() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Submit a result
-        client.submit_result(&0u64, &String::from_str(&env, "abc123"), &Winner::Player1);
-
-        // Read the result
-        let entry = client.get_result(&0u64);
-        assert_eq!(entry.result, Winner::Player1);
-
-        // Verify TTL was extended
-        let ttl = env.as_contract(&contract_id, || {
-            env.storage().persistent().get_ttl(&DataKey::Result(0u64))
-        });
-        assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-    }
-
-    #[test]
-    fn test_pause_twice_is_idempotent() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.pause();
-        client.pause(); // second call must not error
-
-        // Contract is still paused
-        let is_paused: bool = env.as_contract(&contract_id, || {
-            env.storage()
-                .instance()
-                .get(&DataKey::Paused)
-                .unwrap_or(false)
-        });
-        assert!(is_paused);
-    }
-
-    #[test]
-    fn test_unpause_emits_no_event() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // First pause the contract
-        client.pause();
-        // Then unpause it
-        client.unpause();
-
-        // Test passes if unpause completes without panic
-        // The function docstring states it does not emit events
-    }
-
-    #[test]
-    fn test_pause_emits_paused_event() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.pause();
-
-        let events = env.events().all();
-        let expected_topics = soroban_sdk::vec![
-            &env,
-            Symbol::new(&env, "admin").into_val(&env),
-            symbol_short!("paused").into_val(&env),
-        ];
-        let matched = events
-            .iter()
-            .find(|(_, topics, _)| *topics == expected_topics);
-        assert!(matched.is_some(), "paused event not emitted");
-    }
-
-    /// Full integration: oracle stores result, escrow oracle address submits
-    /// result to escrow, payout executes, match is Completed.
-    #[test]
-    fn test_oracle_to_escrow_full_payout_flow() {
-        let (env, oracle_id, escrow_id, _oracle_admin, player1, _player2, token_addr) = setup();
-        let oracle_client = OracleContractClient::new(&env, &oracle_id);
-        let escrow_client = EscrowContractClient::new(&env, &escrow_id);
-        let token_client = soroban_sdk::token::Client::new(&env, &token_addr);
-
-        // Step 1: oracle admin submits verified result to the oracle contract
-        oracle_client.submit_result(
-            &0u64,
-            &String::from_str(&env, "test_game"),
-            &Winner::Player1,
-        );
-        assert!(oracle_client.has_result(&0u64));
-
-        // Step 2: the escrow's trusted oracle address (oracle_admin) calls
-        // submit_result on the escrow contract, triggering the payout
-        escrow_client.submit_result(&0u64, &EscrowWinner::Player1);
-
-        // Step 3: assert match is Completed and player1 received the full pot
-        let m = escrow_client.get_match(&0u64);
-        assert_eq!(m.state, MatchState::Completed);
-        // player1 staked 100, pot = 200; started with 1000, deposited 100 → balance = 1100
-        assert_eq!(token_client.balance(&player1), 1100);
-    }
-
-    #[test]
-    fn test_delete_result_removes_from_storage() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(
-            &0u64,
-            &String::from_str(&env, "chess_game_42"),
-            &Winner::Player1,
-        );
-        assert!(client.has_result(&0u64));
-
-        client.delete_result(&0u64);
-        assert!(!client.has_result(&0u64));
-    }
-
-    #[test]
-    fn test_delete_result_not_found_errors() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        let result = client.try_delete_result(&999u64);
-        assert_eq!(result, Err(Ok(Error::ResultNotFound)));
-    }
-
-    /// Test that delete_result is blocked when the contract is paused.
-    #[test]
-    fn test_delete_result_blocked_when_paused() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        // Submit a result so there is something to delete
-        client.submit_result(
-            &0u64,
-            &String::from_str(&env, "chess_game_99"),
-            &Winner::Player2,
-        );
-        assert!(client.has_result(&0u64));
-
-        // Pause the contract
-        client.pause();
-
-        // Attempt delete_result — must be blocked
-        let result = client.try_delete_result(&0u64);
-        assert_eq!(result, Err(Ok(Error::ContractPaused)));
-
-        // Result must still exist
-        assert!(client.has_result(&0u64));
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_delete_result_requires_admin_auth() {
-        let env = Env::default();
-        // No mock_all_auths — auth is enforced
-        let admin = Address::generate(&env);
-        let contract_id = env.register_contract(None, OracleContract);
-        let client = OracleContractClient::new(&env, &contract_id);
-        client.initialize(&admin);
-        client.delete_result(&0u64);
-    }
-
-    #[test]
-    fn test_instance_ttl_extended_on_submit_result() {
-        let (env, contract_id, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        client.submit_result(&0u64, &String::from_str(&env, "ttl_game"), &Winner::Player1);
-
-        let ttl = env.as_contract(&contract_id, || env.storage().instance().get_ttl());
-        assert_eq!(ttl, crate::MATCH_TTL_LEDGERS);
-    }
-
-    // ── Issue: transfer_admin (update_admin) — old admin rejected, new admin accepted ──
-    //
-    // Verifies that after `update_admin` is called:
-    //   1. The old admin can no longer call `submit_result` (auth failure).
-    //   2. The new admin can successfully call `submit_result`.
-    #[test]
-    fn test_transfer_admin_old_rejected_new_accepted() {
-        let (env, contract_id, _escrow_id, old_admin, _player1, _player2, _token_addr) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        let new_admin = Address::generate(&env);
-
-        // Transfer admin to new_admin (mock_all_auths is active from setup)
-        client.update_admin(&new_admin);
-
-        // ── old admin must be rejected ────────────────────────────────────────
-        // Provide explicit mock auth for old_admin only — new_admin is not mocked,
-        // so the contract will reject old_admin's auth attempt.
-        env.mock_auths(&[soroban_sdk::testutils::MockAuth {
-            address: &old_admin,
-            invoke: &soroban_sdk::testutils::MockAuthInvoke {
-                contract: &contract_id,
-                fn_name: "submit_result",
-                args: (0u64, String::from_str(&env, "test_game"), Winner::Player1).into_val(&env),
-                sub_invokes: &[],
-            },
-        }]);
-
-        let result = client.try_submit_result(
-            &0u64,
-            &String::from_str(&env, "test_game"),
-            &Winner::Player1,
-        );
-        assert!(
-            result.is_err(),
-            "old admin must be rejected after transfer_admin"
-        );
-
-        // ── new admin must be accepted ────────────────────────────────────────
-        env.mock_auths(&[soroban_sdk::testutils::MockAuth {
-            address: &new_admin,
-            invoke: &soroban_sdk::testutils::MockAuthInvoke {
-                contract: &contract_id,
-                fn_name: "submit_result",
-                args: (0u64, String::from_str(&env, "test_game"), Winner::Player1).into_val(&env),
-                sub_invokes: &[],
-            },
-        }]);
-
-        client.submit_result(
-            &0u64,
-            &String::from_str(&env, "test_game"),
-            &Winner::Player1,
-        );
-
-        // Confirm the result was stored
-        assert!(
-            client.has_result(&0u64),
-            "new admin must be able to submit results after transfer"
-        );
-        let entry = client.get_result(&0u64);
-        assert_eq!(entry.result, Winner::Player1);
-    }
-
-    /// Issue #559 — update_admin must emit an admin_rot event with old and new admin.
-    #[test]
-    fn test_update_admin_emits_rotation_event() {
-        let (env, contract_id, _escrow_id, old_admin, ..) = setup();
-        let client = OracleContractClient::new(&env, &contract_id);
-
-        let new_admin = Address::generate(&env);
-        client.update_admin(&new_admin);
-
-        let events = env.events().all();
-        let expected_topics = soroban_sdk::vec![
-            &env,
-            Symbol::new(&env, "admin").into_val(&env),
-            symbol_short!("admin_rot").into_val(&env),
-        ];
-        let matched = events
-            .iter()
-            .find(|(_, topics, _)| *topics == expected_topics);
-        assert!(matched.is_some(), "admin_rot event not emitted");
-
-        let (_, _, data) = matched.unwrap();
-        let (ev_old, ev_new): (Address, Address) =
-            soroban_sdk::TryFromVal::try_from_val(&env, &data).unwrap();
-        assert_eq!(ev_old, old_admin);
-        assert_eq!(ev_new, new_admin);
-    }
-}
